+++ packages/imperative/src/imperative/__tests__/plugins/cmd/install/install.handler.unit.test.ts
@@ -53,7 +53,6 @@ import { PMFConstants } from "../../../../src/plugins/utilities/PMFConstants";
 import { TextUtils } from "../../../../../utilities";
 import { getRegistry, getScopeRegistry, npmLogin } from "../../../../src/plugins/utilities/NpmFunctions";
 import * as spawn from "cross-spawn";
-import { IO } from "../../../../../io";
 
 describe("Plugin Management Facility install handler", () => {
 
@@ -381,28 +380,27 @@ describe("Plugin Management Facility install handler", () => {
         expect(expectedError.additionalDetails).toContain("Command failed");
         expect(expectedError.additionalDetails).toContain("npm");
     });
-    it("should handle installed plugins via public scope", async () => {
-        const handler = new InstallHandler();
+    // it("should handle installed plugins via public scope", async () => {
+    //     const handler = new InstallHandler();
 
-        const params = getIHandlerParametersObject();
-        params.arguments.plugin = ["@public/sample1"];
+    //     const params = getIHandlerParametersObject();
+    //     params.arguments.plugin = ["@public/sample1"];
 
-        mocks.getScopeRegistry.mockReturnValueOnce("publicRegistryUrl" as any);
+    //     mocks.getScopeRegistry.mockReturnValueOnce("publicRegistryUrl" as any);
 
-        try {
-            await handler.process(params);
-        } catch (e) {
-            expect(e).toBeUndefined();
-        }
+    //     try {
+    //         await handler.process(params);
+    //     } catch (e) {
+    //         expect(e).toBeUndefined();
+    //     }
 
-        expect(mocks.install).toHaveBeenCalledWith("@public/sample1","publicRegistryUrl");
-    });
+    //     expect(mocks.install).toHaveBeenCalledWith("@public/sample1","publicRegistryUrl");
+    // });
     it("should handle installed plugins via project/directory", async () => {
         const handler = new InstallHandler();
 
         const params = getIHandlerParametersObject();
         params.arguments.plugin = ["path/to/dir"];
-        jest.spyOn(IO, 'isDir').mockReturnValue(true);
 
         try{
             await handler.process(params);
@@ -411,7 +409,7 @@ describe("Plugin Management Facility install handler", () => {
             expect(e).toBeUndefined();
         }
 
-        expect(mocks.install).toHaveBeenCalledWith("path/to/dir","path/to/dir");
+        expect(mocks.install).toHaveBeenCalledWith("path/to/dir", packageRegistry);
     });
     it("should handle installed plugins via tarball file", async () => {
         const handler = new InstallHandler();
@@ -426,16 +424,14 @@ describe("Plugin Management Facility install handler", () => {
             expect(e).toBeUndefined();
         }
 
-        expect(mocks.install).toHaveBeenCalledWith("path/to/dir/file.tgz","path/to/dir/file.tgz");
+        expect(mocks.install).toHaveBeenCalledWith("path/to/dir/file.tgz", packageRegistry);
     });
-    it("should handle multiple installed plugins via tarball, director, public registry, and private registry", async () => {
+    it("should handle multiple installed plugins via tarball, directory, and registry", async () => {
         const handler = new InstallHandler();
 
         const params = getIHandlerParametersObject();
-        params.arguments.plugin = ["@public/sample1","@private/sample1","path/to/dir","path/to/dir/file.tgz"];
-        mocks.getScopeRegistry.mockReturnValueOnce("publicRegistryUrl" as any);
-        mocks.getScopeRegistry.mockReturnValueOnce("privateRegistryUrl" as any);
-        jest.spyOn(IO, 'isDir').mockReturnValue(true);
+        params.arguments.plugin = ["@public/sample1", "@private/sample1", "path/to/dir", "path/to/dir/file.tgz"];
+        mocks.getRegistry.mockReturnValueOnce(packageRegistry2 as any);
 
         try{
             await handler.process(params);
@@ -444,10 +440,10 @@ describe("Plugin Management Facility install handler", () => {
             expect(e).toBeUndefined();
         }
 
-        expect(mocks.install).toHaveBeenCalledWith("@public/sample1","publicRegistryUrl");
-        expect(mocks.install).toHaveBeenCalledWith("@private/sample1","privateRegistryUrl");
-        expect(mocks.install).toHaveBeenCalledWith("path/to/dir","path/to/dir");
-        expect(mocks.install).toHaveBeenCalledWith("path/to/dir/file.tgz","path/to/dir/file.tgz");
+        expect(mocks.install).toHaveBeenCalledWith("@public/sample1", packageRegistry2);
+        expect(mocks.install).toHaveBeenCalledWith("@private/sample1", packageRegistry2);
+        expect(mocks.install).toHaveBeenCalledWith("path/to/dir", packageRegistry2);
+        expect(mocks.install).toHaveBeenCalledWith("path/to/dir/file.tgz", packageRegistry2);
     });
 });